+++ src/views/Spin.vue
@@ -1,6 +1,21 @@
 <script setup lang="ts">
 import { ref } from 'vue'
 const spinning = ref(true)
+const sizeOptions = [
+  {
+    label: 'small',
+    value: 'small'
+  },
+  {
+    label: 'middle',
+    value: 'middle'
+  },
+  {
+    label: 'large',
+    value: 'large'
+  }
+]
+const size = ref('middle')
 </script>
 <template>
   <div>
@@ -63,33 +78,6 @@ const spinning = ref(true)
         描述文案一起水平垂直居中。
       </p>
     </Spin>
-    <h2 class="mt30 mb10">自定义描述文案</h2>
-    <Space>
-      <Spin class="u-spin" tip="加载中..." :spinning="spinning" />
-      <Spin class="u-spin" tip="加载中..." :spinning="spinning" indicator="spin-dot" />
-      <Spin class="u-spin" tip="加载中..." :spinning="spinning" indicator="spin-line" />
-      <Spin class="u-spin" tip="加载中..." :spinning="spinning" indicator="ring-circle" />
-      <Spin class="u-spin" tip="加载中..." :spinning="spinning" indicator="ring-rail" />
-      <Spin class="u-spin" tip="加载中..." :spinning="spinning" indicator="dynamic-circle" />
-      <Spin class="u-spin" tip="加载中..." :spinning="spinning" indicator="magic-ring" />
-    </Space>
-    <h2 class="mt30 mb10">自定义样式</h2>
-    <Space>
-      <Spin class="u-spin" color="#fadb14" :spinning="spinning" />
-      <Spin class="u-spin" color="#fadb14" :spinning="spinning" indicator="spin-dot" />
-      <Spin class="u-spin" color="#fadb14" :spinning="spinning" indicator="spin-line" />
-      <Spin class="u-spin" color="#fadb14" :spinning="spinning" :spin-circle-percent="50" indicator="ring-circle" />
-      <Spin
-        class="u-spin"
-        color="#fadb14"
-        :spin-circle-percent="50"
-        ring-rail-color="rgba(212, 136, 6, 0.6)"
-        :spinning="spinning"
-        indicator="ring-rail"
-      />
-      <Spin class="u-spin" color="#fadb14" :spinning="spinning" indicator="dynamic-circle" />
-      <Spin class="u-spin" color="#fadb14" magic-ring-color="#ffe58f" :spinning="spinning" indicator="magic-ring" />
-    </Space>
     <h2 class="mt30 mb10">各种大小</h2>
     <Space vertical>
       <Space>
@@ -118,6 +106,54 @@ const spinning = ref(true)
         <Spin class="u-size-spin" :spinning="spinning" size="small" indicator="magic-ring" />
         <Spin class="u-size-spin" :spinning="spinning" indicator="magic-ring" />
         <Spin class="u-size-spin" :spinning="spinning" size="large" indicator="magic-ring" />
+      </Space>
+    </Space>
+    <h2 class="mt30 mb10">自定义描述文案</h2>
+    <Space vertical>
+      <Radio :options="sizeOptions" v-model:value="size" button button-style="solid" />
+      <Space>
+        <Spin class="u-spin" :size="size" tip="加载中..." :spinning="spinning" />
+        <Spin class="u-spin" :size="size" tip="加载中..." :spinning="spinning" indicator="spin-dot" />
+        <Spin class="u-spin" :size="size" tip="加载中..." :spinning="spinning" indicator="spin-line" />
+        <Spin class="u-spin" :size="size" tip="加载中..." :spinning="spinning" indicator="ring-circle" />
+        <Spin class="u-spin" :size="size" tip="加载中..." :spinning="spinning" indicator="ring-rail" />
+        <Spin class="u-spin" :size="size" tip="加载中..." :spinning="spinning" indicator="dynamic-circle" />
+        <Spin class="u-spin" :size="size" tip="加载中..." :spinning="spinning" indicator="magic-ring" />
+      </Space>
+    </Space>
+    <h2 class="mt30 mb10">自定义样式</h2>
+    <Space vertical>
+      <Radio :options="sizeOptions" v-model:value="size" button button-style="solid" />
+      <Space>
+        <Spin class="u-spin" :size="size" color="#fadb14" :spinning="spinning" />
+        <Spin class="u-spin" :size="size" color="#fadb14" :spinning="spinning" indicator="spin-dot" />
+        <Spin class="u-spin" :size="size" color="#fadb14" :spinning="spinning" indicator="spin-line" />
+        <Spin
+          class="u-spin"
+          :size="size"
+          color="#fadb14"
+          :spinning="spinning"
+          :spin-circle-percent="50"
+          indicator="ring-circle"
+        />
+        <Spin
+          class="u-spin"
+          :size="size"
+          color="#fadb14"
+          :spin-circle-percent="50"
+          ring-rail-color="rgba(212, 136, 6, 0.6)"
+          :spinning="spinning"
+          indicator="ring-rail"
+        />
+        <Spin class="u-spin" :size="size" color="#fadb14" :spinning="spinning" indicator="dynamic-circle" />
+        <Spin
+          class="u-spin"
+          :size="size"
+          color="#fadb14"
+          magic-ring-color="#ffe58f"
+          :spinning="spinning"
+          indicator="magic-ring"
+        />
       </Space>
     </Space>
   </div>