+++ docs/guide/changelog.md
@@ -3,6 +3,12 @@
 <BackTop />
 <Watermark fullscreen content="Vue Amazing UI" />
 
+## 1.8.7 <Tag color="volcano" size="small">2024-10-21</Tag>
+
+- 优化并更新 [上传 Upload](https://themusecatcher.github.io/vue-amazing-ui/guide/components/upload.html) 组件，新增 `drop` `preview` 事件回调，调整 `beforeUpload` 属性，支持返回一个 `Promise` 对象，用于控制上传
+- 优化并更新 [加载中 Spin](https://themusecatcher.github.io/vue-amazing-ui/guide/components/upload.html) 组件
+- 组件库及文档代码优化
+
 ## 1.8.6 <Tag color="volcano" size="small">2024-10-18</Tag>
 
 - 优化并更新 [警告提示 Alert](https://themusecatcher.github.io/vue-amazing-ui/guide/components/alert.html) 组件，新增 `type: 'default'` 类型，以及 `bordered` 属性